refactor: split http-server.ts route handlers into domain modules
Extract 1,000+ lines of route handlers from the monolithic http-server.ts
into five focused modules under runtime/routes/:
- conversation-routes.ts (messages, suggestions)
- attachment-routes.ts (upload, delete, get)
- run-routes.ts (create, status, decisions, trust rules)
- channel-routes.ts (inbound, delivery ack, delete conversation)
- app-routes.ts (shareable pages, cloud sharing)

Shared types moved to http-types.ts to avoid circular dependencies.
The main http-server.ts retains only the class shell, router, lifecycle,
health, and interface endpoints (~200 lines).
